Ix: Split unit tests into separate classes
