+++ Ix.NET/Source/System.Interactive.Tests/ReturnTest.cs
@@ -1,0 +1,21 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the Apache 2.0 License.
+// See the LICENSE file in the project root for more information. 
+
+using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Linq;
+using Xunit;
+
+namespace Tests
+{
+    public class ReturnTest : Tests
+    {
+        [Fact]
+        public void Return()
+        {
+            Assert.Equal(42, EnumerableEx.Return(42).Single());
+        }
+    }
+}
