+++ Ix.NET/Source/System.Interactive.Tests/Tests.cs
@@ -9,7 +9,7 @@ using Xunit;
 namespace Tests
 {
 
-    public partial class Tests
+    public class Tests
     {
 #pragma warning disable xUnit1013 // Public method should be marked as test
         public void AssertThrows<E>(Action a)
